Revert generate_training_dataset migration (simplify PR)
only migrate pose estimation, the dataset creation can be addressed in a separate PR
