+++ deeplabcut/generate_training_dataset/multiple_individuals_trainingsetmanipulation.py
@@ -19,25 +19,27 @@ from pathlib import Path
 import numpy as np
 from tqdm import tqdm
 
+import deeplabcut.compat as compat
 import deeplabcut.generate_training_dataset.metadata as metadata
 from deeplabcut.core.config import ProjectConfig
 from deeplabcut.core.engine import Engine
 from deeplabcut.core.weight_init import WeightInitialization
 from deeplabcut.utils import (
+    auxfun_models,
     auxfun_multianimal,
     auxiliaryfunctions,
 )
 
 from .trainingsetmanipulation import (
     MakeInference_yaml,
+    MakeTest_pose_yaml,
+    MakeTrain_pose_yaml,
     SplitTrials,
     merge_annotateddatasets,
     pad_train_test_indices,
     read_image_shape_fast,
     validate_shuffles,
 )
-
-_ENGINE = Engine.PYTORCH
 
 
 def format_multianimal_training_data(
@@ -115,6 +117,7 @@ def create_multianimaltraining_dataset(
     paf_graph_degree=6,
     userfeedback: bool = True,
     weight_init: WeightInitialization | None = None,
+    engine: Engine | None = None,
     ctd_conditions: int | str | Path | tuple[int, str] | tuple[int, int] | None = None,
 ):
     """Creates a training dataset for multi-animal datasets. Labels from all the
@@ -295,6 +298,24 @@ def create_multianimaltraining_dataset(
     if net_type is None:  # loading & linking pretrained models
         net_type = cfg.get("default_net_type", "dlcrnet_ms5")
 
+    # load the engine to use to create the shuffle
+    if engine is None:
+        engine = compat.get_project_engine(cfg)
+
+    if not (any(net in net_type for net in ("resnet", "eff", "dlc", "mob")) or engine == Engine.PYTORCH):
+        raise ValueError(f"Unsupported network {net_type} for engine {engine}.")
+
+    multi_stage = False
+    ### dlcnet_ms5: backbone resnet50 + multi-fusion & multi-stage module
+    ### dlcr101_ms5/dlcr152_ms5: backbone resnet101/152 + multi-fusion & multi-stage module
+    if all(net in net_type for net in ("dlcr", "_ms5")) and engine != Engine.PYTORCH:
+        num_layers = re.findall("dlcr([0-9]*)", net_type)[0]
+        if num_layers == "":
+            num_layers = 50
+        net_type = f"resnet_{num_layers}"
+        multi_stage = True
+
+    dataset_type = "multi-animal-imgaug"
     (
         individuals,
         uniquebodyparts,
@@ -330,10 +351,17 @@ def create_multianimaltraining_dataset(
 
     print("Utilizing the following graph:", partaffinityfield_graph)
     # Disable the prediction of PAFs if the graph is empty
-    bool(partaffinityfield_graph)
+    partaffinityfield_predict = bool(partaffinityfield_graph)
 
     # Loading the encoder (if necessary downloading from TF)
     dlcparent_path = auxiliaryfunctions.get_deeplabcut_path()
+    defaultconfigfile = os.path.join(dlcparent_path, "pose_cfg.yaml")
+
+    if engine == Engine.PYTORCH:
+        model_path = dlcparent_path
+    else:
+        model_path = auxfun_models.check_for_weights(net_type, Path(dlcparent_path))
+
     Shuffles = validate_shuffles(cfg, Shuffles, num_shuffles, userfeedback)
 
     # print(trainIndices,testIndices, Shuffles, augmenter_type,net_type)
@@ -359,7 +387,7 @@ def create_multianimaltraining_dataset(
             splits.append((trainFraction, Shuffles[shuffle], (train_inds, test_inds)))
 
     top_down = False
-    if net_type.startswith("top_down_"):
+    if engine == Engine.PYTORCH and net_type.startswith("top_down_"):
         top_down = True
         net_type = net_type[len("top_down_") :]
 
@@ -401,7 +429,7 @@ def create_multianimaltraining_dataset(
                 cfg=cfg,
                 train_fraction=trainFraction,
                 shuffle=shuffle,
-                engine=_ENGINE,
+                engine=engine,
                 train_indices=trainIndices,
                 test_indices=testIndices,
                 overwrite=not userfeedback,
@@ -423,7 +451,7 @@ def create_multianimaltraining_dataset(
                 trainFraction,
                 shuffle,
                 cfg,
-                engine=_ENGINE,
+                engine=engine,
             )
             auxiliaryfunctions.attempt_to_make_folder(cfg.project_path / modelfoldername, recursive=True)
             auxiliaryfunctions.attempt_to_make_folder(cfg.project_path / modelfoldername / "train")
@@ -433,41 +461,110 @@ def create_multianimaltraining_dataset(
             path_test_config = str(Path(cfg["project_path"]) / modelfoldername / "test" / "pose_cfg.yaml")
             path_inference_config = str(Path(cfg["project_path"]) / modelfoldername / "test" / "inference_cfg.yaml")
 
-            from deeplabcut.pose_estimation_pytorch.config.make_pose_config import (
-                make_pytorch_pose_config,
-                make_pytorch_test_config,
-            )
-            from deeplabcut.pose_estimation_pytorch.modelzoo.config import (
-                make_super_animal_finetune_config,
-            )
+            if engine == Engine.TF:
+                jointnames = [str(bpt) for bpt in multianimalbodyparts]
+                jointnames.extend([str(bpt) for bpt in uniquebodyparts])
+                items2change = {
+                    "dataset": datafilename,
+                    "engine": engine.aliases[0],
+                    "metadataset": metadatafilename,
+                    "num_joints": len(multianimalbodyparts) + len(uniquebodyparts),  # cfg["uniquebodyparts"]),
+                    "all_joints": [
+                        [i] for i in range(len(multianimalbodyparts) + len(uniquebodyparts))
+                    ],  # cfg["uniquebodyparts"]))],
+                    "all_joints_names": jointnames,
+                    "init_weights": str(model_path),
+                    "project_path": str(cfg["project_path"]),
+                    "net_type": net_type,
+                    "multi_stage": multi_stage,
+                    "pairwise_loss_weight": 0.1,
+                    "pafwidth": 20,
+                    "partaffinityfield_graph": partaffinityfield_graph,
+                    "partaffinityfield_predict": partaffinityfield_predict,
+                    "weigh_only_present_joints": False,
+                    "num_limbs": len(partaffinityfield_graph),
+                    "dataset_type": dataset_type,
+                    "optimizer": "adam",
+                    "batch_size": 8,
+                    "multi_step": [[1e-4, 7500], [5 * 1e-5, 12000], [1e-5, 200000]],
+                    "save_iters": 10000,
+                    "display_iters": 500,
+                    "num_idchannel": (len(cfg["individuals"]) if cfg.get("identity", False) else 0),
+                    "crop_size": list(crop_size),
+                    "crop_sampling": crop_sampling,
+                }
 
-            # backwards compatibility with version 2.X
-            if net_type == "dlcrnet_ms5":
-                net_type = "dlcrnet_stride16_ms5"
-
-            config_path = Path(path_train_config).with_name(_ENGINE.pose_cfg_name)
-            if weight_init is not None and weight_init.with_decoder:
-                pytorch_cfg = make_super_animal_finetune_config(
-                    project_config=cfg,
-                    pose_config_path=config_path,
-                    model_name=net_type,
-                    detector_name=detector_type,
-                    weight_init=weight_init,
-                    save=True,
+                trainingdata = MakeTrain_pose_yaml(
+                    items2change,
+                    path_train_config,
+                    defaultconfigfile,
+                    save=(engine == Engine.TF),
                 )
-            else:
-                pytorch_cfg = make_pytorch_pose_config(
-                    project_config=cfg,
-                    pose_config_path=config_path,
-                    net_type=net_type,
-                    top_down=top_down,
-                    detector_type=detector_type,
-                    weight_init=weight_init,
-                    save=True,
-                    ctd_conditions=ctd_conditions,
+                keys2save = [
+                    "dataset",
+                    "num_joints",
+                    "all_joints",
+                    "all_joints_names",
+                    "net_type",
+                    "multi_stage",
+                    "init_weights",
+                    "global_scale",
+                    "location_refinement",
+                    "locref_stdev",
+                    "dataset_type",
+                    "partaffinityfield_predict",
+                    "pairwise_predict",
+                    "partaffinityfield_graph",
+                    "num_limbs",
+                    "dataset_type",
+                    "num_idchannel",
+                ]
+
+                MakeTest_pose_yaml(
+                    trainingdata,
+                    keys2save,
+                    path_test_config,
+                    nmsradius=5.0,
+                    minconfidence=0.01,
+                    sigma=1,
+                    locref_smooth=False,
+                )  # setting important def. values for inference
+            elif engine == Engine.PYTORCH:
+                from deeplabcut.pose_estimation_pytorch.config.make_pose_config import (
+                    make_pytorch_pose_config,
+                    make_pytorch_test_config,
+                )
+                from deeplabcut.pose_estimation_pytorch.modelzoo.config import (
+                    make_super_animal_finetune_config,
                 )
 
-            make_pytorch_test_config(pytorch_cfg, path_test_config, save=True)
+                # backwards compatibility with version 2.X
+                if net_type == "dlcrnet_ms5":
+                    net_type = "dlcrnet_stride16_ms5"
+
+                config_path = Path(path_train_config).with_name(engine.pose_cfg_name)
+                if weight_init is not None and weight_init.with_decoder:
+                    pytorch_cfg = make_super_animal_finetune_config(
+                        project_config=cfg,
+                        pose_config_path=config_path,
+                        model_name=net_type,
+                        detector_name=detector_type,
+                        weight_init=weight_init,
+                        save=True,
+                    )
+                else:
+                    pytorch_cfg = make_pytorch_pose_config(
+                        project_config=cfg,
+                        pose_config_path=config_path,
+                        net_type=net_type,
+                        top_down=top_down,
+                        detector_type=detector_type,
+                        weight_init=weight_init,
+                        save=True,
+                        ctd_conditions=ctd_conditions,
+                    )
+
+                make_pytorch_test_config(pytorch_cfg, path_test_config, save=True)
 
             # Setting inference cfg file:
             default_inf_path = dlcparent_path / "inference_cfg.yaml"
